post: part one of building a profile page with HTML and CSS (#994)
* post: part one of building a profile page with HTML and CSS

This is the first post in the three part series where I go over the process of destructuring a design for a profile page.
We will then dive in and write the code in part two and then conclude the series by taking a brief look at setting up
some testing for the page using Playwright.

* add title to iframe - a11y

---------

Co-authored-by: Schalk Neethling <schalkneethling@Schalks-MacBook-Pro.local>